Extract connected part of app to separate component
